examples/python-todo-app/requirements.txt
grpcio
grpcio-tools
protobuf
hvac
xeze-dbr
